Actualizado el proyecto de visual studio al 2010 Actualizado codigo para compilar en visual studio Es el primer commit en el que wxWidgets compila en windows

+++ src/Cartridge.cpp
@@ -33,7 +33,7 @@ Cartridge::Cartridge(string path)
 	if (file.is_open())
 	{
 		size = file.tellg();
-		this->_RomSize = size;
+		this->_RomSize = (unsigned int)size;
 		_memCartridge = new BYTE [size];
 		file.seekg (0, ios::beg);
 		file.read ((char *)_memCartridge, size);
